+++ crates/itch-source/src/impls/mod.rs
@@ -1,7 +1,11 @@
 //! Default implementations of the `itch-source` traits.
 
+pub mod channel;
+pub mod iterator;
 pub mod null_seq_store;
 pub mod ring_buffer_seq_store;
 
+pub use channel::{ChannelSource, UnboundedChannelSource};
+pub use iterator::IteratorSource;
 pub use null_seq_store::NullSeqStore;
 pub use ring_buffer_seq_store::{RingBufferSeqStore, RingBufferSeqStoreError};